Make Recidiviz Hero capable of bouncing again. Refactor existing code into callable function that is invoked after Hero component is mounted.
