+++ .eslintignore
@@ -1,0 +1,2 @@
+# Ignore static JS files from legacy version of Recidiviz
+**/src/lib/static/*.js
