+++ package.json
@@ -48,7 +48,9 @@
     "webpack-dev-server": "^3.1.3"
   },
   "dependencies": {
+    "animejs": "^2.2.0",
     "babel-runtime": "^6.26.0",
+    "charming": "^2.2.3",
     "prop-types": "^15.6.1",
     "react": "^16.3.2",
     "react-dom": "^16.3.2",
